Do delayed truncations for non-loaded slabs.
This allows us to unmount faster, since we don't need to fetch slabs
at shutdown to truncate them and mark them as removed. Instead we simply
mark the truncation offset in the slabdb, which signals the mgr to truncate
them on their next claim.

Also fix open/slabdb_get ordering in scrub, where I suspect we may have
had a race condition.

We may yet split up the truncation flag in the sqlite DB so we can more
easily query for slabs that need truncation.