Fix inode lock ordering in inode_unload.
The inode lock must never be acquired while holding the open_inodes lock.

Also, the blocks field is intimately tied to the size field. They must
both be protected by the bytes_lock and be written together when
doing a flush for data only.

This reordering of fields means we need to increase ou slab format version.