No longer loop infinitely while claiming slabs. Instead let the error reach the fs layer, where we can then retry until we either get a FUSE interrupt, or if we mark the fs as "offline". When we set "offline" on the filesystem, it will return ENOMEDIUM on any operation that would normally be retried forever (resulting in a 'D' state).
We also adjusted the backend timeout so we don't have to wait too long
for an operation to be retried or interrupted.

All those changes revealed other issues now marked with TODOs. This
feature is not currently stable.

Added a new potatoctl "flush" operation to let users force a sync of
all dirty slabs.