We may not need to keep a reference to open slabs in an inode anymore.
It doesn't get us much. The idea was that directories would always be
preloaded, thus unlikely to fail on a backend timeout. But, that is not
the case since as we modify the directory, new slabs may be needed so
we always need to check for retryable errors on all di_* calls.